// src/app/models/sub-work-order-partnumber.model.ts
export class SubWorkOrderPartNumber {
    constructor() {
        this.promisedDate = new Date();
        this.estimatedShipDate = new Date();
        this.customerRequestDate = new Date();
        this.estimatedCompletionDate = new Date();
        this.customerReference = null;
        this.nte = '';
        this.quantity = 1;
        this.stockLineId = null;
        this.cmmId = 0;
        this.workflowId = 0;
        this.subWorkOrderStageId = 0;
        this.subWorkOrderStatusId = 0;
        this.subWorkOrderPriorityId = 0;
        this.isPMA = false;
        this.isDER = false;
        this.techStationId = 0;
        this.tearDownReport = 0;
        this.tatDaysStandard = 0;
        this.tatDaysCurrent = 0;
        this.technicianId = 0;
        this.mappingItemMasterId = null;
        this.conditionId = null;
        // this.masterPartId = null;
        this.stockLineNumber = '';
        this.isActive = true;
        this.isDelete = false;
        this.updatedDate = new Date();
        this.createdDate = new Date();
        this.receivedDate = '';
        this.isMPNContract=false;
        this.contractNo=null;
        this.workOrderId=0;
        this.workOrderMaterialsId=0;
        this.subWOPartNoId=0;
        this.subWorkOrderScopeId=0;
        this.subWorkOrderId=0;
        this.itemMasterId=0;
        this.masterCompanyId=1;
    }
    masterCompanyId:any;
    itemMasterId:any;
    subWOPartNoId: number;
    subWorkOrderScopeId: number;
    subWorkOrderId: number;
    updatedDate: Date;
    createdDate: Date;
    isActive: boolean
    isDelete: boolean
    stockLineNumber: string;
    // masterPartId: any;
    workOrderScopeId: number;
    promisedDate: any;
    estimatedShipDate: any;
    customerRequestDate: any;
    estimatedCompletionDate: any;
    customerReference: number;
    isMPNContract:boolean;
    contractNo:string;
    nte: string;
    quantity: number;
    stockLineId: number;
    cmmId: number;
    workflowId: number;
    subWorkOrderStageId: number;
    subWorkOrderStatusId: number;
    subWorkOrderPriorityId: number;
    isPMA: boolean;
    isDER: boolean;
    techStationId: number;
    tearDownReport: number;
    tatDaysCurrent: number;
    tatDaysStandard: number;
    technicianId: number;
    mappingItemMasterId: number;
    conditionId: number;
    receivedDate: string;
    workOrderId: any;
    workOrderMaterialsId:any;
}